Move Emotiv handling code to library
